Use Pydantic to handle raw inputs
Fix typing issues

WIP
