+++ src/apify/_actor.py
@@ -2,10 +2,8 @@ from __future__ import annotations
 
 import asyncio
 import os
-import re
 import sys
 from datetime import timedelta
-from itertools import chain
 from typing import TYPE_CHECKING, Any, Callable, TypeVar, cast
 
 from lazy_object_proxy import Proxy
@@ -15,10 +13,8 @@ from typing_extensions import Self
 from apify_client import ApifyClientAsync
 from apify_shared.consts import ActorEnvVars, ActorExitCodes, ApifyEnvVars
 from apify_shared.utils import ignore_docs, maybe_extract_enum_member_value
-from crawlee import Request, service_container
+from crawlee import service_container
 from crawlee.events._types import Event, EventPersistStateData
-from crawlee.http_clients import BaseHttpClient, HttpResponse, HttpxHttpClient
-from crawlee.storages import RequestList
 
 from apify._actor_inputs import _create_request_list
 from apify._configuration import Configuration
@@ -36,13 +32,14 @@ if TYPE_CHECKING:
     import logging
     from types import TracebackType
 
+    from crawlee.http_clients import BaseHttpClient
     from crawlee.proxy_configuration import _NewUrlFunction
+    from crawlee.storages import RequestList
 
     from apify._models import Webhook
 
 
 MainReturnType = TypeVar('MainReturnType')
-
 
 
 class _ActorType:
@@ -982,19 +979,20 @@ class _ActorType:
 
     @staticmethod
     async def create_request_list(
-        *, actor_start_urls_input: list[dict[str,str]], http_client: BaseHttpClient | None= None
+        *, actor_start_urls_input: list[dict[str, Any]], http_client: BaseHttpClient | None = None
     ) -> RequestList:
-        """Creates request list from Actor input requestListSources. This accepts list of urls and requestsFromUrl.
+        """Creates request list from Actor input requestListSources. This accepts list of urls and requests_from_url.
 
         Example:
             actor_start_urls_input = [
                 # Gather urls from response body.
-                {'requestsFromUrl': 'https://crawlee.dev/file.txt', 'method': 'GET'},
+                {'requests_from_url': 'https://crawlee.dev/file.txt', 'method': 'GET'},
                 # Directly include this url.
                 {'url': 'https://crawlee.dev', 'method': 'GET'}
             ]
-            """
+        """
         return await _create_request_list(actor_start_urls_input=actor_start_urls_input, http_client=http_client)
+
 
 Actor = cast(_ActorType, Proxy(_ActorType))
 """The entry point of the SDK, through which all the Actor operations should be done."""
