+++ src/apify/_actor_inputs.py
@@ -1,26 +1,46 @@
+from __future__ import annotations
+
 import asyncio
-from itertools import chain
 import re
+from asyncio import Task
+from typing import Any
+
+from pydantic import BaseModel, Field
 
 from crawlee import Request
-from crawlee.http_clients import BaseHttpClient, HttpxHttpClient, HttpResponse
+from crawlee._types import HttpMethod  #  TODO: Make public in Crawlee?
+from crawlee.http_clients import BaseHttpClient, HttpxHttpClient
 from crawlee.storages import RequestList
 
 URL_NO_COMMAS_REGEX = re.compile(
     r'https?:\/\/(www\.)?([a-zA-Z0-9]|[a-zA-Z0-9][-a-zA-Z0-9@:%._+~#=]{0,254}[a-zA-Z0-9])\.[a-z]{2,63}(:\d{1,5})?(\/[-a-zA-Z0-9@:%_+.~#?&/=()]*)?'
 )
 
+class _RequestDetails(BaseModel):
+    method: HttpMethod
+    payload: str = ''
+    headers: dict[str, str] = Field(default_factory=dict)
+    user_data: dict[str, str]= Field(default_factory=dict, alias='user_data')
+
+class _RequestsFromUrlInput(_RequestDetails):
+    requests_from_url: str = Field(alias='requests_from_url')
+
+class _SimpleUrlInput(_RequestDetails):
+    url: str
+
+
 @staticmethod
 async def _create_request_list(
-    *, actor_start_urls_input: dict, http_client: BaseHttpClient | None = None
+    *, actor_start_urls_input: list[dict[str, Any]], http_client: BaseHttpClient | None = None
 ) -> RequestList:
     if not http_client:
         http_client = HttpxHttpClient()
     simple_url_requests_inputs = [
-        request_input for request_input in actor_start_urls_input if 'url' in request_input
-    ]
+        _SimpleUrlInput(**request_input) for request_input in actor_start_urls_input
+        if 'url' in request_input]
     remote_url_requests_inputs = [
-        request_input for request_input in actor_start_urls_input if 'requestsFromUrl' in request_input
+        _RequestsFromUrlInput(**request_input) for request_input in actor_start_urls_input
+        if 'requests_from_url' in request_input
     ]
 
     simple_url_requests = _create_requests_from_input(simple_url_requests_inputs)
@@ -30,44 +50,41 @@ async def _create_request_list(
 
 
 @staticmethod
-def _create_requests_from_input(simple_url_requests_inputs: list[dict[str, str]]) -> list[Request]:
+def _create_requests_from_input(simple_url_inputs: list[_SimpleUrlInput]) -> list[Request]:
     return [
         Request.from_url(
-            method=request_input.get('method'),
-            url=request_input.get('url'),
-            payload=request_input.get('payload', '').encode('utf-8'),
-            headers=request_input.get('headers', {}),
-            user_data=request_input.get('userData', {}),
+            method=request_input.method,
+            url=request_input.url,
+            payload=request_input.payload.encode('utf-8'),
+            headers=request_input.headers,
+            user_data=request_input.user_data,
         )
-        for request_input in simple_url_requests_inputs
+        for request_input in simple_url_inputs
     ]
 
 
 @staticmethod
 async def _create_requests_from_url(
-    remote_url_requests_inputs: list[dict[str, str]], http_client: BaseHttpClient
+    remote_url_requests_inputs: list[_RequestsFromUrlInput], http_client: BaseHttpClient
 ) -> list[Request]:
+    created_requests: list[Request] = []
+
+    def extract_requests_from_response(task: Task) -> list[Request]:
+        matches = re.finditer(URL_NO_COMMAS_REGEX, task.result().read().decode('utf-8'))
+        created_requests.extend([Request.from_url(match.group(0)) for match in matches])
+
     remote_url_requests = []
     for request_input in remote_url_requests_inputs:
-        remote_url_requests.append(
-            asyncio.create_task(
-                http_client.send_request(
-                    method=request_input['method'],
-                    url=request_input['requestsFromUrl'],
-                    headers=request_input.get('headers', {}),
-                    payload=request_input.get('payload', '').encode('utf-8'),
-                )
+        task = asyncio.create_task(
+            http_client.send_request(
+                method=request_input.method,
+                url=request_input.requests_from_url,
+                headers=request_input.headers,
+                payload=request_input.payload.encode('utf-8'),
             )
         )
+        task.add_done_callback(extract_requests_from_response)
+        remote_url_requests.append(task)
+
     await asyncio.gather(*remote_url_requests)
-    # TODO as callbacks
-    a = list(
-        extract_requests_from_response(finished_request.result()) for finished_request in remote_url_requests
-    )
-    return list(chain.from_iterable(a))
-
-
-@staticmethod
-def extract_requests_from_response(response: HttpResponse) -> list[Request]:
-    matches = list(re.finditer(URL_NO_COMMAS_REGEX, response.read().decode('utf-8')))
-    return [Request.from_url(match.group(0)) for match in matches]
+    return created_requests
